% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_count_targets.R
\name{find_and_count_targets}
\alias{find_and_count_targets}
\title{find  and count possible targets in an image}
\usage{
find_and_count_targets(paths, keep1 = 50)
}
\arguments{
\item{paths}{a data frame with the full paths to the images in a column named "path".}

\item{keep1}{the number of high-MH-distance 16x16 pixel squares to keep in the first pass.}
}
\value{
A dataframe with counts of targets in each image
}
\description{
find  and count possible targets in an image
}
