% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_crop_targets.R
\name{find_and_crop_targets}
\alias{find_and_crop_targets}
\title{find possible targets and crop out little squares around them}
\usage{
find_and_crop_targets(img, keep1 = 200, keep2 = 100, outdir = ".",
  crop_pixels = 150, max_cumul_overlap = 0.2)
}
\arguments{
\item{img}{the path to the image.}

\item{keep1}{the number of high-MH-distance 16x16 pixel squares to keep in the first pass.}

\item{keep2}{the number of squares that actually gets kept to be reported.
The actual number may be less than this because overlapping squares get tossed.}

\item{outdir}{the directory in which to write the results}

\item{crop_pixels}{how many pixels on a side in what gets cropped out?}

\item{max_cumul_overlap}{if a square overlaps in area cumulatively by more than this amount with any other squares with
better targets, it is tossed}
}
\description{
Plots a 100 x 100 sheet of possible targets and a figure showing where
they occur.
}
