% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_fraction.R
\name{overlap_fraction_single}
\alias{overlap_fraction_single}
\title{Given hi and lo x coordinates of two squares, return the fraction of the second squares's area that overlaps the first}
\usage{
overlap_fraction_single(xlo1, xhi1, ylo1, yhi1, xlo2, xhi2, ylo2, yhi2,
  use1)
}
\arguments{
\item{xlo1}{the low x value of square 1, etc...}

\item{use1}{a logical vector.  If TRUE then the corresponding square one should
be considered and used to contribute to any overlap.}
}
\description{
The parameter here should be fairly obvious.
This is implemented vectorized on square 1.  So the square 1 parameters can
be vectors
}
\keyword{internal}
